Break the original script into 2
